planning-poker-backend/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.6
pydantic==2.11.3
pydantic-settings==2.8.1
email-validator==2.2.0
pwdlib[argon2]==0.2.1
PyJWT==2.10.1
